Fixed Github Issue #4
- Fixed issue where a custom build and load path entered in an asset groups' BundledAssetGroupSchema would result in several errors during the build process. This is an issue in vanilla Addressables still.
- Fixed issue where a leader '/'-character was placed in paths for external catalog bundles.
- Updated the external catalogs to also adopt the ProfileValueReference variable types for its build and load paths.
- Updated documentation.
- Updated the catalog generation to also support binary formatted catalogs.

+++ Editor/Build/DataBuilders/BuildScriptPackedMultiCatalogMode.cs
@@ -1,7 +1,9 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using UnityEditor.AddressableAssets.Settings;
+using UnityEditor.AddressableAssets.Settings.GroupSchemas;
 using UnityEngine;
 using UnityEngine.AddressableAssets;
 using UnityEngine.AddressableAssets.Initialization;
@@ -23,6 +25,11 @@ namespace UnityEditor.AddressableAssets.Build.DataBuilders
 		/// <param name="dst">the destination</param>
 		private static void FileMoveOverwrite(string src, string dst)
 		{
+			if (src == dst)
+			{
+				return;
+			}
+
 			if (File.Exists(dst))
 			{
 				File.Delete(dst);
@@ -57,25 +64,48 @@ namespace UnityEditor.AddressableAssets.Build.DataBuilders
 			{
 				if (catalogContentGroup != null)
 				{
-					catalogSetups.Add(new CatalogSetup(catalogContentGroup));
+					catalogSetups.Add(new CatalogSetup(catalogContentGroup, builderInput));
 				}
 			}
 
 			// Assign assets to new catalogs based on included groups
 			var profileSettings = aaContext.Settings.profileSettings;
 			var profileId = aaContext.Settings.activeProfileId;
+			var defaultLoadPathData = profileSettings.GetProfileDataByName(AddressableAssetSettings.kLocalLoadPath);
+
 			foreach (var loc in aaContext.locations)
 			{
-				CatalogSetup preferredCatalog = catalogSetups.FirstOrDefault(cs => cs.CatalogContentGroup.IsPartOfCatalog(loc, aaContext));
+				var preferredCatalog = catalogSetups.FirstOrDefault(cs => cs.CatalogContentGroup.IsPartOfCatalog(loc, aaContext));
 				if (preferredCatalog != null)
 				{
 					if (loc.ResourceType == typeof(IAssetBundleResource))
 					{
-						string filePath = Path.GetFullPath(loc.InternalId.Replace("{UnityEngine.AddressableAssets.Addressables.RuntimePath}", Addressables.BuildPath));
-						string runtimeLoadPath = preferredCatalog.CatalogContentGroup.RuntimeLoadPath + "/" + Path.GetFileName(filePath);
-						runtimeLoadPath = profileSettings.EvaluateString(profileId, runtimeLoadPath);
+						var bundleId = (loc.Data as AssetBundleRequestOptions).BundleName + ".bundle";
+						var group = aaContext.Settings.FindGroup(g => (g != null) && (g.Guid == aaContext.bundleToAssetGroup[bundleId]));
 
-						preferredCatalog.Files.Add(filePath);
+						if (group == null)
+						{
+							Debug.LogErrorFormat("Could not find the group that belongs to location {0}.", loc.InternalId);
+							continue;
+						}
+
+						// If the group's load path is set to be the global local load path, then it is switched out for the one defined by the external catalog.
+						var schema = group.GetSchema<BundledAssetGroupSchema>();
+						var loadPath = (schema.LoadPath.Id != defaultLoadPathData.Id) ? schema.LoadPath : preferredCatalog.CatalogContentGroup.RuntimeLoadPath;
+
+						// Evaluate the load path, and if it amounts to nothing, then pick the Id itself as the load path, as it might be a custom one.
+						var runtimeLoadPath = loadPath.GetValue(profileSettings, profileId);
+						if (string.IsNullOrEmpty(runtimeLoadPath))
+						{
+							runtimeLoadPath = profileSettings.EvaluateString(profileId, loadPath.Id);
+						}
+
+						if (!string.IsNullOrEmpty(runtimeLoadPath) && !runtimeLoadPath.EndsWith('/'))
+						{
+							runtimeLoadPath += '/';
+						}
+
+						runtimeLoadPath += Path.GetFileName(loc.InternalId);
 						preferredCatalog.BuildInfo.Locations.Add(new ContentCatalogDataEntry(typeof(IAssetBundleResource), runtimeLoadPath, loc.Provider, loc.Keys, loc.Dependencies, loc.Data));
 					}
 					else
@@ -89,27 +119,29 @@ namespace UnityEditor.AddressableAssets.Build.DataBuilders
 				}
 			}
 
-
-			// Process dependencies
 			foreach (CatalogSetup additionalCatalog in catalogSetups)
 			{
-				var dataEntries = new Queue<ContentCatalogDataEntry>(additionalCatalog.BuildInfo.Locations);
-				var processedEntries = new HashSet<ContentCatalogDataEntry>();
-				while (dataEntries.Count > 0)
+				var locationQueue = new Queue<ContentCatalogDataEntry>(additionalCatalog.BuildInfo.Locations);
+				var processedLocations = new HashSet<ContentCatalogDataEntry>();
+
+				while (locationQueue.Count > 0)
 				{
-					ContentCatalogDataEntry dataEntry = dataEntries.Dequeue();
-					if (!processedEntries.Add(dataEntry) || (dataEntry.Dependencies == null) || (dataEntry.Dependencies.Count == 0))
+					ContentCatalogDataEntry location = locationQueue.Dequeue();
+
+					// If the location has already been processed, or doesn't have any dependencies, then skip it.
+					if (!processedLocations.Add(location) || (location.Dependencies == null) || (location.Dependencies.Count == 0))
 					{
 						continue;
 					}
 
-					foreach (var entryDependency in dataEntry.Dependencies)
+					foreach (var entryDependency in location.Dependencies)
 					{
 						// Search for the dependencies in the default catalog only.
 						var depLocation = defaultCatalog.Locations.Find(loc => loc.Keys[0] == entryDependency);
+
 						if (depLocation != null)
 						{
-							dataEntries.Enqueue(depLocation);
+							locationQueue.Enqueue(depLocation);
 
 							// If the dependency wasn't part of the catalog yet, add it.
 							if (!additionalCatalog.BuildInfo.Locations.Contains(depLocation))
@@ -135,6 +167,7 @@ namespace UnityEditor.AddressableAssets.Build.DataBuilders
 					catalogs.Add(setup.BuildInfo);
 				}
 			}
+
 			return catalogs;
 		}
 
@@ -143,22 +176,70 @@ namespace UnityEditor.AddressableAssets.Build.DataBuilders
 			// execute build script
 			var result = base.DoBuild<TResult>(builderInput, aaContext);
 
-			// move extra catalogs to CatalogsBuildPath
+			// Move out the extra catalogs and their content.
 			foreach (var setup in catalogSetups)
 			{
-				// Empty catalog setups are not added/built
+				// Empty catalog setups are not added/built.
 				if (setup.Empty)
 				{
 					continue;
 				}
 
-				var bundlePath = aaContext.Settings.profileSettings.EvaluateString(aaContext.Settings.activeProfileId, setup.CatalogContentGroup.BuildPath);
-				Directory.CreateDirectory(bundlePath);
+				var builtInShaderBundleName = GetBuiltInShaderBundleNamePrefix(aaContext) + "_unitybuiltinshaders";
+				var profileSettings = aaContext.Settings.profileSettings;
+				var activeProfileId = aaContext.Settings.activeProfileId;
+				var defaultBuildPathData = profileSettings.GetProfileDataByName(AddressableAssetSettings.kLocalBuildPath);
+				var globalBuildPath = profileSettings.EvaluateString(activeProfileId, profileSettings.GetValueById(activeProfileId, defaultBuildPathData.Id));
 
-				FileMoveOverwrite(Path.Combine(Addressables.BuildPath, setup.BuildInfo.JsonFilename), Path.Combine(bundlePath, setup.BuildInfo.JsonFilename));
-				foreach (var file in setup.Files)
+				// Move the catalog to its destination.
+				var catalogBuildPath = setup.CatalogContentGroup.BuildPath.GetValue(profileSettings, activeProfileId);
+				if (string.IsNullOrEmpty(catalogBuildPath))
 				{
-					FileMoveOverwrite(file, Path.Combine(bundlePath, Path.GetFileName(file)));
+					catalogBuildPath = setup.CatalogContentGroup.BuildPath.Id;
+
+					if (string.IsNullOrWhiteSpace(catalogBuildPath))
+					{
+						throw new Exception("The catalog build path for external catalog " + setup.CatalogContentGroup.name + " is empty.");
+					}
+				}
+
+				Directory.CreateDirectory(catalogBuildPath);
+				FileMoveOverwrite(Path.Combine(Addressables.BuildPath, setup.BuildInfo.CatalogFilename), Path.Combine(catalogBuildPath, setup.BuildInfo.CatalogFilename));
+
+				foreach (var loc in setup.BuildInfo.Locations)
+				{
+					if (loc.ResourceType != typeof(IAssetBundleResource))
+					{
+						continue;
+					}
+
+					var bundleRequestOptions = loc.Data as AssetBundleRequestOptions;
+
+					// Skip the built-in shader collection. That one needs to remain in the main catalog.
+					// This one may end up in here because it is a dependency of a lot of location data entries.
+					if (bundleRequestOptions.BundleName.Equals(builtInShaderBundleName))
+					{
+						continue;
+					}
+
+					var bundleId = bundleRequestOptions.BundleName + ".bundle";
+					var group = aaContext.Settings.FindGroup(g => (g != null) && (g.Guid == aaContext.bundleToAssetGroup[bundleId]));
+					var schema = group.GetSchema<BundledAssetGroupSchema>();
+
+					// If the schema defines a different build path than the default catalog's build path,
+					// then the user made a conscious action to place the bundle somewhere and nothing else has
+					// to be done anymore. If it's still pointing towards the default catalog's build path,
+					// then it should be moved out.
+					if (schema.BuildPath.Id != defaultBuildPathData.Id)
+					{
+						continue;
+					}
+
+					// Move the bundle out of the default build location.
+					var bundleName = Path.GetFileName(loc.InternalId);
+					var bundleSrcPath = Path.Combine(globalBuildPath, bundleName);
+					var bundleDstPath = Path.Combine(catalogBuildPath, bundleName);
+					FileMoveOverwrite(bundleSrcPath, bundleDstPath);
 				}
 			}
 
@@ -175,15 +256,34 @@ namespace UnityEditor.AddressableAssets.Build.DataBuilders
 			}
 
 			// Cleanup the additional catalogs
-			AddressableAssetSettings settings = AddressableAssetSettingsDefaultObject.Settings;
-			foreach (ExternalCatalogSetup additionalCatalog in externalCatalogs)
+			var profileSettings = AddressableAssetSettingsDefaultObject.Settings.profileSettings;
+			var profileId = AddressableAssetSettingsDefaultObject.Settings.activeProfileId;
+
+			var libraryDirectory = new DirectoryInfo("Library");
+			var assetsDirectory = new DirectoryInfo("Assets");
+
+			foreach (ExternalCatalogSetup externalCatalog in externalCatalogs)
 			{
-				string buildPath = settings.profileSettings.EvaluateString(settings.activeProfileId, additionalCatalog.BuildPath);
+				string buildPath = externalCatalog.BuildPath.GetValue(profileSettings, profileId);
+				if (string.IsNullOrEmpty(buildPath))
+				{
+					buildPath = externalCatalog.BuildPath.Id;
+				}
+
 				if (!Directory.Exists(buildPath))
 				{
 					continue;
 				}
 
+				// Stop if we're about to delete the whole library or assets directory.
+				var buildDirectory = new DirectoryInfo(buildPath);
+				if ((Path.GetRelativePath(buildDirectory.FullName, libraryDirectory.FullName) == ".") ||
+					(Path.GetRelativePath(buildDirectory.FullName, assetsDirectory.FullName) == "."))
+				{
+					continue;
+				}
+
+				// Delete each file in the build directory.
 				foreach (string catalogFile in Directory.GetFiles(buildPath))
 				{
 					File.Delete(catalogFile);
@@ -200,25 +300,21 @@ namespace UnityEditor.AddressableAssets.Build.DataBuilders
 			/// <summary>
 			/// The catalog build info.
 			/// </summary>
-			public readonly ContentCatalogBuildInfo BuildInfo;
-
-			/// <summary>
-			/// The files associated to the catalog.
-			/// </summary>
-			public readonly List<string> Files = new List<string>(1);
+			public readonly ContentCatalogBuildInfo BuildInfo = null;
 
 			/// <summary>
 			/// Tells whether the catalog is empty.
 			/// </summary>
 			public bool Empty
 			{
-				get { return BuildInfo.Locations.Count == 0; }
+				get => BuildInfo.Locations.Count == 0;
 			}
 
-			public CatalogSetup(ExternalCatalogSetup buildCatalog)
+			public CatalogSetup(ExternalCatalogSetup buildCatalog, AddressablesDataBuilderInput builderInput)
 			{
 				this.CatalogContentGroup = buildCatalog;
-				BuildInfo = new ContentCatalogBuildInfo(buildCatalog.CatalogName, buildCatalog.CatalogName + ".json");
+				string catalogFileName = string.Format("{0}{1}", buildCatalog.CatalogName, Path.GetExtension(builderInput.RuntimeCatalogFilename));
+				BuildInfo = new ContentCatalogBuildInfo(buildCatalog.CatalogName, catalogFileName);
 				BuildInfo.Register = false;
 			}
 		}
